refactor: polished Tabs with helper handleTabKeyboardNavigation
